try to speed up optimized version of aom
